refactor: rename componet
